+++ src/System/Support/Facades/Schedule.php
@@ -8,7 +8,8 @@ namespace System\Support\Facades;
  * @method static \System\Cron\ScheduleTime[] getPools()
  * @method static \System\Cron\ScheduleTime   call(\Closure $call_back, array $params = [])
  * @method static void                        execute()
- * @method static void                        setLogger()
+ * @method static void                        setLogger(\System\Cron\Schedule\InterpolateInterface $logger)
+ * @method static void                        setTime(int $time)
  * @method static \System\Cron\Schedule       add(\System\Cron\Schedule $schedule)
  * @method static void                        flush()
  */